feat: implement training orchestration and validation utilities for SBERT training
